# api-server/api/openapi.yaml
openapi: 3.0.3
info:
  title: MassBank Tool API
  version: "0.1"
servers:
- url: https://localhost:8080/
paths:
  /spectra:
    get:
      operationId: getAllSpectra
      parameters:
      - description: number of items
        explode: true
        in: query
        name: limit
        required: false
        schema:
          format: int64
          type: integer
        style: form
      - description: the offset of items
        explode: true
        in: query
        name: offset
        required: false
        schema:
          format: int64
          type: integer
        style: form
      - description: the page number
        explode: true
        in: query
        name: page
        required: false
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getAllSpectra_200_response'
          description: A list of spectra
  /spectrainfo:
    get:
      operationId: getAllSpectraInfo
      parameters:
      - description: number of items
        explode: true
        in: query
        name: limit
        required: false
        schema:
          format: int64
          type: integer
        style: form
      - description: the offset of items
        explode: true
        in: query
        name: offset
        required: false
        schema:
          format: int64
          type: integer
        style: form
      - description: the page number
        explode: true
        in: query
        name: page
        required: false
        schema:
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getAllSpectra_200_response'
          description: A list of spectra
  /spectra/{accession}:
    get:
      operationId: getSpectrum
      parameters:
      - description: an accession id
        explode: false
        in: path
        name: accession
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Spectrum'
          description: OK
        "404":
          description: Not Found
  /upload/massbank:
    post:
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/_upload_massbank_post_request'
      responses:
        "200":
          description: OK
        "409":
          description: Could not insert dataset
  /upload/mzml:
    post:
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/_upload_massbank_post_request'
      responses:
        "200":
          description: OK
        "409":
          description: Could not insert dataset
  /spectra/rebuildgit:
    post:
      responses:
        "200":
          description: OK
components:
  parameters:
    offsetParam:
      description: the offset of items
      explode: true
      in: query
      name: offset
      required: false
      schema:
        format: int64
        type: integer
      style: form
    limitParam:
      description: number of items
      explode: true
      in: query
      name: limit
      required: false
      schema:
        format: int64
        type: integer
      style: form
    pageParam:
      description: the page number
      explode: true
      in: query
      name: page
      required: false
      schema:
        format: int64
        type: integer
      style: form
    accession:
      description: an accession id
      explode: false
      in: path
      name: accession
      required: true
      schema:
        type: string
      style: simple
  schemas:
    SpectrumListItem:
      additionalProperties:
        type: string
      example:
        accession: accession
      properties:
        accession:
          type: string
      required:
      - accession
      type: object
    Spectrum:
      additionalProperties:
        type: string
      example:
        accession: accession
      properties:
        accession:
          type: string
      required:
      - accession
      type: object
    getAllSpectra_200_response:
      example:
        spectra:
        - accession: accession
        - accession: accession
        totalCount: 0
      properties:
        totalCount:
          type: integer
        spectra:
          items:
            $ref: '#/components/schemas/SpectrumListItem'
          type: array
      type: object
    _upload_massbank_post_request:
      properties:
        filename:
          type: string
        file:
          format: binary
          type: string
      type: object